chore(team): split channel manager tests
